services/logic/requirements.txt
aio-pika==8.2.0
aiohttp==3.8.1
aiormq==6.4.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.82.0
frozenlist==1.3.1
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
multidict==6.0.2
pamqp==3.2.0
pydantic==1.10.2
pypika-tortoise==0.1.6
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.19.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
yarl==1.8.1
